refactor expiry lambda
