+++ handlers/digital-subscription-expiry/src/main/scala/com/gu/digitalSubscriptionExpiry/DigitalSubscriptionExpirySteps.scala
@@ -1,0 +1,51 @@
+package com.gu.digitalSubscriptionExpiry
+
+import com.gu.util.Logging
+import com.gu.util.apigateway.ApiGatewayHandler.Operation
+import com.gu.util.apigateway.ResponseModels.{ApiResponse, Headers}
+import com.gu.util.apigateway.{ApiGatewayRequest}
+import com.gu.util.reader.Types._
+import main.scala.com.gu.digitalSubscriptionExpiry.DigitalSubscriptionExpiryRequest
+import org.joda.time.DateTime
+import play.api.libs.json.{Json}
+
+import scalaz.{-\/}
+
+object DigitalSubscriptionExpirySteps extends Logging {
+
+  def getZuoraExpiry() = DigitalSubscriptionExpiryResponse(Expiry(
+    expiryDate = DateTime.now(),
+    expiryType = ExpiryType.SUB,
+    subscriptionCode = None,
+    provider = Some("test provider")
+  ))
+
+  def getEmergencyTokenExpiry() = {
+
+  }
+  def apply(): Operation = {
+
+    def steps(apiGatewayRequest: ApiGatewayRequest): FailableOp[Unit] = {
+      val calloutParsed: Option[DigitalSubscriptionExpiryRequest] = Json.fromJson[DigitalSubscriptionExpiryRequest](Json.parse(apiGatewayRequest.body)).asOpt
+
+      logger.info(s"Parsed request as: $calloutParsed")
+
+      val responseJson = Json.toJson(getZuoraExpiry())
+
+      -\/(ApiResponse("200", new Headers, Json.prettyPrint(responseJson)))
+    }
+
+    //TODO
+    //    def healthcheck() =
+    //      for {
+    //        identityId <- getByEmail(EmailAddress("john.duffell@guardian.co.uk")).leftMap(a => ApiGatewayResponse.internalServerError(a.toString)).withLogging("healthcheck getByEmail")
+    //        _ <- countZuoraAccountsForIdentityId(identityId)
+    //        _ <- sfAuth()
+    //      } yield ()
+
+    Operation.noHealthcheck(steps, false)
+
+  }
+
+}
+
